// src/store/modules/tabs/action-types.ts
export enum TabsActionTypes {
	ACTION_ADD_VIEW = 'ACTION_ADD_VIEW',
	ACTION_ADD_VISITED_VIEW = 'ACTION_ADD_VISITED_VIEW',
	ACTION_ADD_CACHED_VIEW = 'ACTION_ADD_CACHED_VIEW',

	ACTION_DEL_VIEW = 'ACTION_DEL_VIEW',
	ACTION_DEL_VISITED_VIEW = 'ACTION_DEL_VISITED_VIEW',
	ACTION_DEL_CACHED_VIEW = 'ACTION_DEL_CACHED_VIEW',

	ACTION_DEL_OTHERS_VIEWS = 'ACTION_DEL_OTHERS_VIEWS',
	ACTION_DEL_OTHERS_VISITED_VIEWS = 'ACTION_DEL_OTHERS_VISITED_VIEWS',
	ACTION_DEL_OTHERS_CACHED_VIEWS = 'ACTION_DEL_OTHERS_CACHED_VIEWS',

	ACTION_DEL_ALL_VIEWS = 'ACTION_DEL_ALL_VIEWS',
	ACTION_DEL_ALL_VISITED_VIEWS = 'ACTION_DEL_ALL_VISITED_VIEWS',
	ACTION_DEL_ALL_CACHED_VIEWS = 'ACTION_DEL_ALL_CACHED_VIEWS',

	ACTION_UPDATE_VISITED_VIEW = 'ACTION_UPDATE_VISITED_VIEW'
}