// src/store/modules/tabs/mutation-types.ts
export enum TabsMutationTypes {
	SET_ADD_VISITED_VIEW = 'SET_ADD_VISITED_VIEW',
	SET_ADD_CACHED_VIEW = 'SET_ADD_CACHED_VIEW',
	SET_DEL_VISITED_VIEW = 'SET_DEL_VISITED_VIEW',
	SET_DEL_CACHED_VIEW = 'SET_DEL_CACHED_VIEW',
	SET_DEL_OTHERS_VISITED_VIEWS = 'SET_DEL_OTHERS_VISITED_VIEWS',
	SET_DEL_OTHERS_CACHED_VIEWS = 'SET_DEL_OTHERS_CACHED_VIEWS',
	SET_DEL_ALL_VISITED_VIEWS = 'SET_DEL_ALL_VISITED_VIEWS',
	SET_DEL_ALL_CACHED_VIEWS = 'SET_DEL_ALL_CACHED_VIEWS',
	SET_UPDATE_VISITED_VIEW = 'SET_UPDATE_VISITED_VIEW'
}